refactor(cli): extract launchGoalSession from session start (#276)
